Make adjoint tape interpreter eager
